# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HSD Auth Platform API
  description: |
    HSD Auth Platform - Authentication as a Service API
    
    Bu API, HSD projelerinde kimlik doğrulama işlemleri için kullanılır.
    
    ## Authentication
    API istekleri için JWT Bearer token veya API Key kullanılabilir.
    
    ## Rate Limiting
    - Varsayılan: 100 istek/dakika
    - Rate limit aşıldığında 429 Too Many Requests döner
  version: 1.0.0
  contact:
    name: HSD Auth Support
    email: support@hsdcore.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.auth.hsdcore.com
    description: Production server
  - url: https://api.staging.auth.hsdcore.com
    description: Staging server
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /admin/realms:
    get:
      tags:
        - Realms
      summary: Realm listesi
      operationId: listRealms
      responses:
        '200':
          description: Realm listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Realm'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Realms
      summary: Yeni realm oluştur
      operationId: createRealm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRealmRequest'
      responses:
        '201':
          description: Realm oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Realm'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /admin/realms/{realmId}/users:
    get:
      tags:
        - Admin
      summary: Realm kullanıcılarını listele
      operationId: listRealmUsers
      parameters:
        - $ref: '#/components/parameters/RealmId'
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: status
          in: query
          schema:
            type: string
            enum: [active, disabled, pending]
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Kullanıcı listesi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedUsers'
    post:
      tags:
        - Admin
      summary: Kullanıcı oluştur
      operationId: createUser
      parameters:
        - $ref: '#/components/parameters/RealmId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Kullanıcı oluşturuldu
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    RealmId:
      name: realmId
      in: path
      required: true
      schema:
        type: string
      description: Realm ID
...
  responses:
    Unauthorized:
      description: Kimlik doğrulama gerekli
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Yetkisiz erişim
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    ValidationError:
      description: Doğrulama hatası
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
    ValidationError:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        emailVerified:
          type: boolean
        status:
          type: string
          enum: [active, disabled, pending]
        profile:
          $ref: '#/components/schemas/Profile'
        roles:
          type: array
          items:
            type: string
        mfaEnabled:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Profile:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
        metadata:
          type: object
    ProfileInput:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: string
...
    Realm:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, disabled]
        settings:
          $ref: '#/components/schemas/RealmSettings'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RealmSettings:
      type: object
      properties:
        sessionTimeout:
          type: integer
          description: Oturum zaman aşımı (saniye)
        maxConcurrentSessions:
          type: integer
        passwordPolicy:
          $ref: '#/components/schemas/PasswordPolicy'
        mfa:
          $ref: '#/components/schemas/MfaSettings'
    PasswordPolicy:
      type: object
      properties:
        minLength:
          type: integer
        requireUppercase:
          type: boolean
        requireLowercase:
          type: boolean
        requireNumbers:
          type: boolean
        requireSpecialChars:
          type: boolean
    MfaSettings:
      type: object
      properties:
        enabled:
          type: boolean
        required:
          type: boolean
        methods:
          type: array
          items:
            type: string
            enum: [totp, sms]
    CreateRealmRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        displayName:
          type: string
        description:
          type: string
        settings:
          $ref: '#/components/schemas/RealmSettings'
...
    CreateUserRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        profile:
          $ref: '#/components/schemas/ProfileInput'
        roles:
          type: array
          items:
            type: string
        emailVerified:
          type: boolean
        requirePasswordReset:
          type: boolean
...
    PaginatedUsers:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/User'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer